pyproject.toml
[project]
name = "horizon-min"
version = "0.1.0"
description = "Duck Hunt VLM Environment - A headless game environment for training and evaluating Vision Language Models"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [
    { name = "David Mayboroda" }
]
keywords = ["vlm", "reinforcement-learning", "game-environment", "duck-hunt", "ai"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "pillow>=10.0.0",
    "fastapi>=0.115.0",
    "uvicorn>=0.32.0",
    "pydantic>=2.0.0",
    "requests>=2.28.0",
]

[project.optional-dependencies]
training = [
    "torch>=2.4.0",
    "transformers>=4.45.0",
    "trl>=0.15.0",
    "peft>=0.13.0",
    "accelerate>=1.0.0",
    "datasets>=3.0.0",
    "bitsandbytes>=0.44.0",
    "pyyaml>=6.0",
    "wandb>=0.18.0",
    "huggingface_hub>=0.25.0",
    "openai>=1.0.0",
    # flash-attn is optional — install separately if needed:
    #   uv pip install flash-attn --no-build-isolation
    # Without it, set model.attn_implementation to "sdpa" or "eager"
]
experiments = [
    "weave>=0.51.0",
    "openai>=1.0.0",
    "wandb>=0.18.0",
]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "ruff>=0.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["duck_hunt_openenv"]

[tool.uv]
dev-dependencies = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "ruff>=0.1.0",
]

[tool.ruff]
target-version = "py310"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "W"]
ignore = ["E501"]

[tool.pytest.ini_options]
testpaths = ["duck_hunt_openenv/tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = "-v"
